src/golang.org/x/net/html/testdata/webkit/tests21.dat

#data
<svg><![CDATA[foo]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<math><![CDATA[foo]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       "foo"

#data
<div><![CDATA[foo]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[foo
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[foo
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]] >]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]] >]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]]
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]]"

#data
<svg><![CDATA[]
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]"

#data
<svg><![CDATA[]>q
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]>q"

#data
<svg><foreignObject><div><![CDATA[foo]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <div>
|           <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[<svg>]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[</svg>q]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>q"

#data
<svg><![CDATA[<svg>q
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>q"

#data
<svg><![CDATA[</svg>q
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>q"

#data
<svg><![CDATA[<svg>]]><path>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <svg path>

#data
<svg><![CDATA[<svg>]]></path>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[<svg>]]><!--path-->
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <!-- path -->

#data
<svg><![CDATA[<svg>]]>path
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>path"

#data
<svg><![CDATA[<!--svg-->]]>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<!--svg-->"
